refactor: move plugin integration tests to new module
