+++ gravitee-apim-integration-tests/src/test/java/io/gravitee/apim/integration/tests/messages/HttpPostEntrypointKafkaEndpointIntegrationTest.java
@@ -1,4 +1,4 @@
-package io.gravitee.plugin.tests.messages;/**
+package io.gravitee.apim.integration.tests.messages;/**
  * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -16,39 +16,43 @@ package io.gravitee.plugin.tests.messages;/**
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.graviteesouce.reactor.message.MessageApiReactorFactory;
 import io.gravitee.apim.gateway.tests.sdk.AbstractGatewayTest;
 import io.gravitee.apim.gateway.tests.sdk.annotations.DeployApi;
 import io.gravitee.apim.gateway.tests.sdk.annotations.GatewayTest;
 import io.gravitee.apim.gateway.tests.sdk.connector.EndpointBuilder;
 import io.gravitee.apim.gateway.tests.sdk.connector.EntrypointBuilder;
-import io.gravitee.common.http.MediaType;
+import io.gravitee.apim.gateway.tests.sdk.reactor.ReactorBuilder;
+import io.gravitee.apim.plugin.reactor.ReactorPlugin;
 import io.gravitee.definition.model.v4.Api;
-import io.gravitee.gateway.api.http.HttpHeaderNames;
+import io.gravitee.gateway.reactive.reactor.v4.reactor.ReactorFactory;
 import io.gravitee.gateway.reactor.ReactableApi;
 import io.gravitee.plugin.endpoint.EndpointConnectorPlugin;
 import io.gravitee.plugin.endpoint.kafka.KafkaEndpointConnectorFactory;
 import io.gravitee.plugin.entrypoint.EntrypointConnectorPlugin;
-import io.gravitee.plugin.entrypoint.http.get.HttpGetEntrypointConnectorFactory;
+import io.gravitee.plugin.entrypoint.http.post.HttpPostEntrypointConnectorFactory;
 import io.vertx.core.http.HttpMethod;
-import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
-import io.vertx.kafka.client.producer.RecordMetadata;
+import io.vertx.kafka.client.common.TopicPartition;
 import io.vertx.rxjava3.core.Vertx;
 import io.vertx.rxjava3.core.http.HttpClient;
-import io.vertx.rxjava3.kafka.client.producer.KafkaProducer;
-import io.vertx.rxjava3.kafka.client.producer.KafkaProducerRecord;
+import io.vertx.rxjava3.kafka.client.consumer.KafkaConsumer;
+import io.vertx.rxjava3.kafka.client.producer.KafkaHeader;
 import java.util.Collection;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
+import java.util.UUID;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 import org.apache.kafka.clients.admin.AdminClient;
 import org.apache.kafka.clients.admin.AdminClientConfig;
 import org.apache.kafka.clients.admin.NewTopic;
-import org.apache.kafka.clients.producer.ProducerConfig;
-import org.apache.kafka.common.serialization.ByteArraySerializer;
-import org.apache.kafka.common.serialization.StringSerializer;
+import org.apache.kafka.clients.consumer.ConsumerConfig;
+import org.apache.kafka.common.serialization.ByteArrayDeserializer;
+import org.apache.kafka.common.serialization.StringDeserializer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
@@ -65,14 +69,18 @@ import org.testcontainers.utility.DockerImageName;
  */
 @Testcontainers
 @GatewayTest
-@DeployApi({ "/apis/v4/messages/http-get-entrypoint-kafka-endpoint.json" })
-@Disabled
-class HttpGetEntrypointKafkaEndpointIntegrationTest extends AbstractGatewayTest {
+@DeployApi({ "/apis/v4/messages/http-post-entrypoint-kafka-endpoint.json" })
+class HttpPostEntrypointKafkaEndpointIntegrationTest extends AbstractGatewayTest {
 
     @Container
     private static final KafkaContainer kafka = new KafkaContainer(DockerImageName.parse("confluentinc/cp-kafka:6.2.1"));
 
     public static final String TEST_TOPIC = "test-topic";
+
+    @Override
+    public void configureReactors(Set<ReactorPlugin<? extends ReactorFactory<?>>> reactors) {
+        reactors.add(ReactorBuilder.build(MessageApiReactorFactory.class));
+    }
 
     @Override
     public void configureEndpoints(Map<String, EndpointConnectorPlugin<?, ?>> endpoints) {
@@ -81,7 +89,7 @@ class HttpGetEntrypointKafkaEndpointIntegrationTest extends AbstractGatewayTest 
 
     @Override
     public void configureEntrypoints(Map<String, EntrypointConnectorPlugin<?, ?>> entrypoints) {
-        entrypoints.putIfAbsent("http-get", EntrypointBuilder.build("http-get", HttpGetEntrypointConnectorFactory.class));
+        entrypoints.putIfAbsent("http-post", EntrypointBuilder.build("http-post", HttpPostEntrypointConnectorFactory.class));
     }
 
     @Override
@@ -112,62 +120,66 @@ class HttpGetEntrypointKafkaEndpointIntegrationTest extends AbstractGatewayTest 
     }
 
     @Test
-    @DisplayName("Should deploy a V4 API with a HTTP Get entrypoint and Kafka endpoint")
-    void shouldBeAbleToSubscribeToKafkaEndpointWithHTTPGetEntrypoint(HttpClient client, Vertx vertx) {
-        // In order to simplify the test, Kafka endpoint's consumer is configured with "autoOffsetReset": "earliest"
-        // It allows us to publish the messages in the topic before opening the api connection through SSE entrypoint.
-        KafkaProducer<String, byte[]> producer = getKafkaProducer(vertx);
-        publishMessage(producer, "message1");
-        publishMessage(producer, "message2");
-        publishMessage(producer, "message3");
-        producer.close();
+    @DisplayName("Should deploy a V4 API with a HTTP Post entrypoint and Kafka endpoint")
+    void shouldBeAbleToSubscribeToKafkaEndpointWithHTTPPostEntrypoint(HttpClient client, Vertx vertx) {
+        JsonObject requestBody = new JsonObject();
+        requestBody.put("field", "value");
 
         client
-            .rxRequest(HttpMethod.GET, "/test")
+            .rxRequest(HttpMethod.POST, "/test")
             .flatMap(request -> {
-                request.putHeader(HttpHeaderNames.ACCEPT.toString(), MediaType.APPLICATION_JSON);
-                return request.rxSend();
+                request.putHeader("X-Test-Header", "header-value");
+                return request.rxSend(requestBody.toString());
             })
-            .flatMap(response -> {
-                assertThat(response.statusCode()).isEqualTo(200);
-                return response.body();
+            .flatMapPublisher(response -> {
+                assertThat(response.statusCode()).isEqualTo(202);
+                return response.toFlowable();
             })
             .test()
             .awaitDone(10, TimeUnit.SECONDS)
-            .assertValue(body -> {
-                final JsonObject jsonResponse = new JsonObject(body.toString());
-                final JsonArray items = jsonResponse.getJsonArray("items");
-                assertThat(items).hasSize(2);
-                final JsonObject message = items.getJsonObject(0);
-                assertThat(message.getString("content")).isEqualTo("message1");
-                final JsonObject message2 = items.getJsonObject(1);
-                assertThat(message2.getString("content")).isEqualTo("message2");
-                return true;
-            })
+            .assertComplete()
+            .assertNoErrors();
+
+        // Configure a KafkaConsumer to read messages published on topic test-topic.
+        KafkaConsumer<String, byte[]> kafkaConsumer = getKafkaConsumer(vertx);
+        TopicPartition topicPartition = new TopicPartition(TEST_TOPIC, 0);
+
+        kafkaConsumer
+            .rxAssign(topicPartition)
+            .andThen(kafkaConsumer.rxSeekToBeginning(topicPartition))
+            .andThen(kafkaConsumer.toFlowable())
+            // We expect one message for this test
+            .take(1)
+            .test()
+            .awaitDone(30, TimeUnit.SECONDS)
+            .assertValueAt(
+                0,
+                message -> {
+                    assertThat(message.headers()).contains(KafkaHeader.header("X-Test-Header", "header-value"));
+                    final io.vertx.kafka.client.consumer.KafkaConsumerRecord kafkaConsumerRecord = message.getDelegate();
+                    assertThat(kafkaConsumerRecord.value()).isEqualTo(requestBody.toBuffer().getBytes());
+                    return true;
+                }
+            )
             .assertComplete();
+
+        kafkaConsumer.close();
     }
 
     /**
-     * Creates a KafkaProducer to be able to publish messages to topic
+     * Creates a KafkaConsumer to be able to read messages from topic
      * @param vertx
      * @return
      */
-    private static KafkaProducer<String, byte[]> getKafkaProducer(Vertx vertx) {
-        Map<String, String> config = Map.of(
-            ProducerConfig.BOOTSTRAP_SERVERS_CONFIG,
-            kafka.getBootstrapServers(),
-            ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG,
-            StringSerializer.class.getName(),
-            ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG,
-            ByteArraySerializer.class.getName()
-        );
-
-        return KafkaProducer.create(vertx, config);
-    }
-
-    private static RecordMetadata publishMessage(KafkaProducer<String, byte[]> producer, String message1) {
-        return producer
-            .rxSend(KafkaProducerRecord.create(TEST_TOPIC, "key", io.gravitee.gateway.api.buffer.Buffer.buffer(message1).getBytes()))
-            .blockingGet();
+    private static KafkaConsumer<String, byte[]> getKafkaConsumer(Vertx vertx) {
+        Map<String, String> config = new HashMap<>();
+        config.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, kafka.getBootstrapServers());
+        config.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class.getName());
+        config.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, ByteArrayDeserializer.class.getName());
+        config.put(ConsumerConfig.GROUP_ID_CONFIG, UUID.randomUUID().toString());
+        config.put(ConsumerConfig.CLIENT_ID_CONFIG, UUID.randomUUID().toString());
+        config.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest");
+        KafkaConsumer<String, byte[]> kafkaConsumer = KafkaConsumer.create(vertx, config);
+        return kafkaConsumer;
     }
 }
