+++ gravitee-apim-integration-tests/src/test/resources/apis/v4/messages/http-get-entrypoint-kafka-endpoint.json
