+++ gravitee-apim-integration-tests/src/test/resources/apis/v4/messages/http-post-entrypoint-kafka-endpoint.json
