+++ gravitee-apim-integration-tests/src/test/resources/apis/v4/messages/sse-entrypoint-kafka-endpoint.json
